requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
imbalanced-learn>=0.11.0
tensorflow>=2.14.0
keras-tuner>=1.4.0
matplotlib>=3.7.0
PyYAML>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
lime>=0.2.0.1
